requirements.txt
factory_boy>=2.5.2,<3.0.0
Flask==0.12
mysqlclient==1.3.9
SQLAlchemy==1.1.5
ujson==1.35